cpufreq: exynos-acme: Fix build warning
drivers/cpufreq/exynos-acme.c:1185:22:
warning: comparison of distinct pointer types ('typeof (domain->max_freq) *' (aka 'unsigned int *')
and 'typeof (freq_table[index]) *' (aka 'unsigned long *')) [-Wcompare-distinct-pointer-types]

Bug: 173452474

+++ drivers/cpufreq/exynos-acme.c
@@ -1182,7 +1182,7 @@ static int init_domain(struct exynos_cpufreq_domain *domain,
 			if (volt_table[index] <= val)
 				break;
 		}
-		domain->max_freq = min(domain->max_freq, freq_table[index]);
+		domain->max_freq = min(domain->max_freq, (unsigned int)freq_table[index]);
 		domain->max_freq_qos = domain->max_freq;
 	}
